requirements.txt
pandas
requests
beautifulsoup4
scikit-learn
streamlit
plotly
pyyaml
